+++ library/symbol_table.h
@@ -23,7 +23,7 @@ public:
     void PutBottom(const std::string& key, ValueType value);
     const ValueType& Get(const std::string& key) const;
     const ValueType& GetTopOnly(const std::string& key) const;
-    int depth() const { return this->map_vector_.size(); };
+    size_t depth() const { return this->map_vector_.size(); };
 
 private:
     std::vector< std::unordered_map<std::string, ValueType, std::hash<std::string> > > map_vector_;
